Streamlined FFIBackends, common code to base abstract class
