+++ hat/backends/ffi/cuda/src/main/java/hat/backend/ffi/CudaBackend.java
@@ -25,10 +25,7 @@
 package hat.backend.ffi;
 
 
-import hat.ComputeContext;
 import hat.Config;
-import hat.NDRange;
-import hat.buffer.DispatchContext;
 import hat.callgraph.KernelCallGraph;
 import hat.callgraph.MethodCallDag;
 import jdk.incubator.code.CodeTransformer;
@@ -48,7 +45,6 @@ import jdk.incubator.code.dialect.core.SSA;
 
 import java.lang.foreign.Arena;
 import java.lang.invoke.MethodHandles;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -57,6 +53,25 @@ import static optkl.OpHelper.Invoke;
 import static optkl.OpHelper.Invoke.invoke;
 
 public class CudaBackend extends C99FFIBackend {
+    public CudaBackend(Config config) {
+        super(Arena.global(), MethodHandles.lookup(),"cuda_backend", config);
+    }
+
+    public CudaBackend() {
+        this(Config.fromEnvOrProperty());
+    }
+
+
+    @Override public String createCode(KernelCallGraph kernelCallGraph,  Object... justArgs){
+        if (config().ptx()){
+            return createPTX(kernelCallGraph,  justArgs);
+        }else{
+            return createCode(kernelCallGraph, new CudaHATKernelBuilder(kernelCallGraph,new ScopedCodeBuilderContext(kernelCallGraph.lookup(),kernelCallGraph.callDag.entryPoint.funcOp())), justArgs);
+        }
+    }
+
+    // The rest of this is needed only for PTX.
+
     final int major = 7;
     final int minor = 5;
     final String target = "sm_52";
@@ -365,45 +380,7 @@ public class CudaBackend extends C99FFIBackend {
 
     final Set<String> usedMathFns = new HashSet<>();
 
-    public CudaBackend(Config config) {
-        super(Arena.global(), MethodHandles.lookup(),"cuda_backend", config);
-    }
 
-    public CudaBackend() {
-        this(Config.fromEnvOrProperty());
-    }
-    @Override
-    public void computeContextHandoff(ComputeContext computeContext) {
-        computeContext.computeCallGraph().callDag.entryPoint.funcOp(injectBufferTracking(config(),lookup(),computeContext.computeCallGraph().callDag.entryPoint.funcOp()));
-    }
-
-    @Override
-    public void dispatchKernel(KernelCallGraph kernelCallGraph, NDRange ndRange, Object... dispatchContextAndArgs) {
-        if (!(dispatchContextAndArgs[0] instanceof DispatchContext )){
-            throw new RuntimeException("No dispatch context");
-        }
-        Object[] justArgs = Arrays.copyOfRange(dispatchContextAndArgs,1,dispatchContextAndArgs.length);
-        CompiledKernel compiledKernel = kernelCallGraphCompiledCodeMap.computeIfAbsent(kernelCallGraph, (_) -> {
-            String code =config().ptx() ? createPTX(kernelCallGraph,  justArgs) : createC99(kernelCallGraph, justArgs);
-            if (config().showCode()) {
-                System.out.println(code);
-            }
-            var compilationUnit = backendBridge.compile(code);
-            if (compilationUnit.ok()) {
-                var kernel = compilationUnit.getKernel(kernelCallGraph.callDag.entryPoint.method().getName());
-                return new CompiledKernel(this, kernelCallGraph,  kernel, dispatchContextAndArgs);
-            } else {
-                throw new IllegalStateException("cuda failed to compile ");
-            }
-        });
-        compiledKernel.dispatch(ndRange, dispatchContextAndArgs);
-    }
-
-    String createC99(KernelCallGraph kernelCallGraph, Object... args){
-        return createCode(kernelCallGraph, new CudaHATKernelBuilder(kernelCallGraph,new ScopedCodeBuilderContext(kernelCallGraph.lookup(),kernelCallGraph.callDag.entryPoint.funcOp())), args);
-    }
-
-    ///   Same as OpenCL backend until here
 
 
     String createPTX(KernelCallGraph kernelCallGraph,  Object... args){
