+++ hat/backends/ffi/opencl/src/main/java/hat/backend/ffi/OpenCLBackend.java
@@ -23,17 +23,12 @@
  * questions.
  */
 package hat.backend.ffi;
-
-import hat.ComputeContext;
 import hat.Config;
-import hat.NDRange;
-import hat.buffer.DispatchContext;
 import hat.callgraph.KernelCallGraph;
 import optkl.codebuilders.ScopedCodeBuilderContext;
 
 import java.lang.foreign.Arena;
 import java.lang.invoke.MethodHandles;
-import java.util.Arrays;
 
 public class OpenCLBackend extends C99FFIBackend {
     public OpenCLBackend(Config config) {
@@ -45,36 +40,7 @@ public class OpenCLBackend extends C99FFIBackend {
     }
 
     @Override
-    public void computeContextHandoff(ComputeContext computeContext) {
-        computeContext.computeCallGraph().callDag.entryPoint.funcOp(injectBufferTracking(config(), lookup(), computeContext.computeCallGraph().callDag.entryPoint.funcOp()));
+    public String createCode(KernelCallGraph kernelCallGraph, Object... justArgs) {
+        return createCode(kernelCallGraph, new OpenCLHATKernelBuilder(kernelCallGraph, new ScopedCodeBuilderContext(kernelCallGraph.lookup(), kernelCallGraph.callDag.entryPoint.funcOp())), justArgs);
     }
-
-    @Override
-    final public void dispatchKernel(KernelCallGraph kernelCallGraph,  NDRange ndRange, Object... dispatchContextAndArgs) {
-        if (!(dispatchContextAndArgs[0] instanceof DispatchContext )){
-            throw new RuntimeException("No dispatch context");
-        }
-        Object[] justArgs = Arrays.copyOfRange(dispatchContextAndArgs,1,dispatchContextAndArgs.length);
-        CompiledKernel compiledKernel = kernelCallGraphCompiledCodeMap.computeIfAbsent(kernelCallGraph, (_) -> {
-            String code = createC99(kernelCallGraph, justArgs);
-            if (config().showCode()) {
-                System.out.println(code);
-            }
-            var compilationUnit = backendBridge.compile(code);
-            if (compilationUnit.ok()) {
-                var kernel = compilationUnit.getKernel(kernelCallGraph.callDag.entryPoint.method().getName());
-                return new CompiledKernel(this, kernelCallGraph, kernel, dispatchContextAndArgs);
-            } else {
-                // TODO: We should capture the log from OpenCL and provide as exception message
-                throw new IllegalStateException("OpenCL program failed to compile");
-            }
-        });
-        compiledKernel.dispatch(ndRange, dispatchContextAndArgs);
-    }
-
-
-    String createC99(KernelCallGraph kernelCallGraph, Object[] args) {
-        return createCode(kernelCallGraph, new OpenCLHATKernelBuilder(kernelCallGraph, new ScopedCodeBuilderContext(kernelCallGraph.lookup(), kernelCallGraph.callDag.entryPoint.funcOp())), args);
-    }
-
 }
