--- hat/backends/ffi/shared/src/main/java/hat/backend/ffi/FFIBackend.java
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.  Oracle designates this
- * particular file as subject to the "Classpath" exception as provided
- * by Oracle in the LICENSE file that accompanied this code.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-
-package hat.backend.ffi;
-
-import hat.ComputeContext;
-import hat.Config;
-import hat.KernelContext;
-
-import java.lang.foreign.Arena;
-import java.lang.invoke.MethodHandles;
-
-public abstract class FFIBackend extends FFIBackendDriver {
-
-    public FFIBackend(Arena arena, MethodHandles.Lookup lookup, String libName, Config config) {
-        super(arena, lookup, libName, config);
-    }
-
-    public void dispatchCompute(ComputeContext computeContext, Object... args) {
-
-        backendBridge.computeStart();
-        // if (config().interpret()) {
-        //   computeContext.interpretWithArgs(args);
-        // } else {
-        if (args[0].equals(KernelContext.class)) {
-            computeContext.invokeWithArgs(args);
-        }else{
-            computeContext.newInvokeWithArgs(args);
-        }
-
-        //}
-        backendBridge.computeEnd();
-    }
-}
